Move source compilation to a build step, add feature detects at runtime to reduce repetition in compiled sources file
