+++ lib/index.js
@@ -8,6 +8,9 @@ if (!String.prototype.startsWith) {
 	require('../polyfills/String.prototype.startsWith/polyfill');
 }
 
+require('es6-promise').polyfill();
+
+
 // Load additional useragent features: primarily to use: agent.satisfies to
 // test a browser version against a semver string
 require('useragent/features');
@@ -137,9 +140,14 @@ function getPolyfillString(options) {
 
 			graph.add(featureName);
 			var polyfill = sourceslib.getPolyfill(featureName).variants[feature.polyfillVariant];
-			var srcKey = (options.minify ? 'min':'raw') + ((feature.flags.indexOf('gated') !== -1) ? 'Gated':'') + 'Source';
+			var srcKey = (options.minify ? 'min':'raw') + 'Source';
+
 
 			feature.polyfillOutput = polyfill[srcKey];
+
+			if (feature.flags.indexOf('gated') !== -1 && polyfill.detectSource) {
+				feature.polyfillOutput = "if (!(" + polyfill.detectSource + ")) {" + feature.polyfillOutput + "}\n\n";
+			}
 
 			if (polyfill.dependencies) {
 				for (var i = 0; i < polyfill.dependencies.length; i++) {
